+++ hdt-java-core/src/main/java/org/rdfhdt/hdt/compact/sequence/SequenceInt64.java
@@ -168,7 +168,7 @@ public class SequenceInt64 implements DynamicSequence {
 	@Override
 	public void add(Iterator<Long> iterator) {
 		while (iterator.hasNext()) {
-			long value = iterator.next().longValue();
+			long value = iterator.next();
 			append(value);
 		}
 	}
